Updated Website & Pages
Fixed the website according to what Professor Perry wanted.
(Made the "Our Team" page the Homepage and edited it to what she wanted, made the Final Report page with placeholder PDF, made Final Report Presentation Video page with placeholder video, and overall made the website nicer to look at.)

All that needs to be added is the Final Report and Final Report Presentation Video when those are finished to replace the placeholders in the code.